+++ spec/factories/books.rb
@@ -12,6 +12,7 @@ class Book < ActiveRecord::Base
        key book_index_by_bin(bin, id)
     )
     EOF
+    Book.delete_all
   end
 
 end
@@ -27,10 +28,17 @@ end
 class BookTemp < ActiveRecord::Base
 
   self.table_name = 'book_vault'
+  self.primary_key= 'book_id'
 
   def self.create_table
     connection.execute <<-EOF
-    create temporary table if not exists book_vault like books
+    create temporary table if not exists
+    book_vault (
+       `book_id` int(11) primary key auto_increment,
+       `bin` int(11),
+       `published_by` varchar(64)
+    )
     EOF
+    BookTemp.delete_all
   end
 end