+++ CHANGES.md
@@ -1,5 +1,10 @@
 See the [documentation](http://bkayser.github.io/cleansweep) for details
 
-# Version 1.0.1
+### Version 1.0.1
 
 * Initial release
+
+### Version 1.0.2
+
+* Changed destination options so you can delete from a different table.
+* Added `dest_columns` option as a map of column names in the source to column names in the destination.